refactor(agent/prompts): extract tests from mod.rs to sibling
Moves the inline #[cfg(test)] mod tests block out of prompts/mod.rs into
prompts/tests.rs to bring the production file under the project's
<=500 line guideline and reduce mod.rs weight per the "light mod.rs" rule.
No behavior change.
